RTB-2383: Resolved L2 comments
